refac: Move codes unrelated to the Owner to packages/webauthn
1. Move unrelated passkeyOwnerWebAuthn/{background, content, tabs}
   code to packages/webauthn/ for better code logic expression.
