+++ packages/webAuthn/content/method.ts
